# Add performance test for handle.py / BinaryValue