--- a/tests/test_cases/test_perf_handle/README.md
+++ b/tests/test_cases/test_perf_handle/README.md
@@ -1,0 +1,6 @@
+# Cocotb Performance test
+
+Benchmarking different ways of reading / writing signals in cocotb
+
+When accessing signals, cocotb uses BinaryValue, this benchmark allows us to compare different ways
+of accessing signals.