These tests are measuring how long it takes to read signals, read one specific bit, write signals in cocotb.

This is part of the "Replace BinaryValue with LogicArray" effort https://github.com/cocotb/cocotb/pull/3244, these tests will help us better compare different implementations and their pros/cons.

Example of output:
```
100025000.00ns INFO     cocotb.dff8                        Little Endian, setting input signal: 3.1270687309988716s
200035000.00ns INFO     cocotb.dff8                        Big Endian, setting input signal: 3.1491839480004273s
300045000.00ns INFO     cocotb.dff8                        Little Endian, reading the output signal: 2.52074205900135
400055000.00ns INFO     cocotb.dff8                        Big Endian, reading the output signal: 2.66596318899974
500065000.00ns INFO     cocotb.dff8                        Little Endian, reading part of the output signal: 2.780234088000725
600075000.00ns INFO     cocotb.dff8                        Big Endian Endian, reading part of the output sign: 2.776298863000193
700085000.00ns INFO     cocotb.dff8                        Little Endian, reading single bit from the output signal: 2.6840617939997173
800095000.00ns INFO     cocotb.dff8                        Big Endian, reading single bit from the output signal: 2.7572237929998664
800115000.00ns INFO     cocotb.dff8                        Little Endian, setting part of the input signal: Unsupported
800125000.00ns INFO     cocotb.dff8                        Big Endian, setting part of the input signal: Unsupported
900125000.00ns INFO     cocotb.dff8                        Little Endian, setting single bit of the input signal: 3.229350190000332
1000135000.00ns INFO     cocotb.dff8                       Big Endian, setting single bit of the input signal: 3.3048068640000565
1100155000.00ns INFO     cocotb.dff8                       Little Endian, read + write signals: 3.469675129999814
1200175000.00ns INFO     cocotb.dff8                       Big Endian, read + write signals: 3.5726902460010024
```