@ktbarrett 2023-09-17
It's probably easier to gather performance metrics using pytest-benchmark and the test_binary_value.py in the pytest directory.