@aubindetrez 2023-09-17
The benchmark introduced by this PR is meant to be implementation independent.
`test_binary_value.py` only tests BinaryValue, however I would like an end-to-end test that does not depend on any specific implementation.
I would use it to check how different implementations of LogicArray perform as a replacement for BinaryValue.

What alternative do we have (that would ideally not require a simulator to run)?
I considered a `pytest` for `handle.py`, but how much of `handle.py` require a simulator to run anyway?