@themperek 2023-10-20
Example benchmark code is here: https://github.com/cocotb/cocotb/blob/master/tests/benchmark.py
It is running in CI: https://github.com/cocotb/cocotb/actions/workflows/benchmark.yml

It needs fixing (authentication settings) to actually export and make plots.