@aubindetrez 2023-10-28
Thanks, I have already been using this benchmark for my BinaryValue vs LogicArray comparison.

But I would also need to benchmark other aspects of the BinaryValue/LogicArray API, like what if a signal contains `X` or `Z`, or if we only access a few bits?
Dedicated benchmarks would be very helpful to compare different implementations and optimization as we are doing refactoring...