@themperek 2023-10-28
>But I would also need to benchmark other aspects of the BinaryValue/LogicArray API, like what if a signal contains X or Z, or if we only access a few bits?

Feel free to add. 

I will try to figure out how to get plotting/authentication working.