@ktbarrett 2025-09-29
We have the matrix_multiplier running in a performance benchmark: https://cocotb.github.io/cocotb-benchmark-results/dev/bench/